thêm tab thông báo

+++ app/src/main/java/com/lhd/fragment/LichThiFragment.java
@@ -227,7 +227,6 @@ public class LichThiFragment extends Fragment {
             SimpleDateFormat timeFormat= new SimpleDateFormat("hh:mm:ss dd/MM/yyyy");
             String s=timeFormat.format(today.getTime());
             s=s.split(" ")[1];
-
             String ngay1=s.split("/")[0];
             String thang1=s.split("/")[1];
             String nam1=s.split("/")[2];
@@ -240,7 +239,7 @@ public class LichThiFragment extends Fragment {
                 if ((Double.parseDouble(ngay2)-Double.parseDouble(ngay1))<0)
                     toi="Đã thi";
                 else
-                toi="Còn lại "+(Double.parseDouble(ngay2)-Double.parseDouble(ngay1))+ " ngày để ôn :)";
+                toi="Còn lại "+(int)(Double.parseDouble(ngay2)-Double.parseDouble(ngay1))+ " ngày để ôn :)";
             }else if (Double.parseDouble(thang1)<Double.parseDouble(thang2) &&nam1.equals(nam2)){
                 toi="Chuẩn bị thi :(";
             }else
